Refactor Inspector into modular tabs and utils
